Move `Read.ProtocolParameters` to `Write.ProtocolParameters`.
This commit moves the recently-introduced `ProtocolParameters` module
from the `Read` module hierarchy to the `Write` module hierarchy.
