+++ lib/wallet/api/http/Cardano/Wallet/Api/Http/Shelley/Server.hs
@@ -681,8 +681,8 @@ import qualified Cardano.Wallet.Primitive.Types.UTxO as UTxO
 import qualified Cardano.Wallet.Primitive.Types.UTxOIndex.Internal as UTxOIndex
 import qualified Cardano.Wallet.Primitive.Types.UTxOSelection as UTxOSelection
 import qualified Cardano.Wallet.Read as Read
-import qualified Cardano.Wallet.Read.ProtocolParameters as Read
 import qualified Cardano.Wallet.Registry as Registry
+import qualified Cardano.Wallet.Write.ProtocolParameters as Write
 import qualified Cardano.Wallet.Write.Tx as WriteTx
 import qualified Cardano.Wallet.Write.Tx.Balance as Write
 import qualified Control.Concurrent.Concierge as Concierge
@@ -3062,7 +3062,7 @@ balanceTransaction
                         txLayer
                         genInpScripts
                         mScriptTemplate)
-                    (Read.ProtocolParameters pp nodePParams)
+                    (Write.ProtocolParameters pp nodePParams)
                     ti
                     utxoIndex
                     (W.defaultChangeAddressGen argGenChange (Proxy @k))
