+++ lib/wallet/src/Cardano/Wallet.hs
@@ -635,7 +635,7 @@ import qualified Cardano.Wallet.Primitive.Types.UTxOIndex as UTxOIndex
 import qualified Cardano.Wallet.Primitive.Types.UTxOSelection as UTxOSelection
 import qualified Cardano.Wallet.Primitive.Types.UTxOStatistics as UTxOStatistics
 import qualified Cardano.Wallet.Read as Read
-import qualified Cardano.Wallet.Read.ProtocolParameters as Read
+import qualified Cardano.Wallet.Write.ProtocolParameters as Write
 import qualified Cardano.Wallet.Write.Tx as WriteTx
 import qualified Cardano.Wallet.Write.Tx.Balance as Write
 import qualified Data.ByteArray as BA
@@ -2104,7 +2104,7 @@ buildAndSignTransactionPure
         (unsignedBalancedTx, updatedWalletState) <- lift $
             buildTransactionPure @s @k @n @recentEra
                 wallet ti utxoIndex txLayer changeAddrGen
-                (Read.unsafeFromWalletProtocolParameters protocolParams)
+                (Write.unsafeFromWalletProtocolParameters protocolParams)
                 preSelection txCtx
         put wallet { getState = updatedWalletState }
 
@@ -2203,7 +2203,7 @@ buildTransaction DBLayer{..} txLayer timeInterpreter walletId
                 utxoIndex
                 txLayer
                 changeAddrGen
-                (Read.unsafeFromWalletProtocolParameters protocolParameters)
+                (Write.unsafeFromWalletProtocolParameters protocolParameters)
                 PreSelection { outputs = paymentOuts }
                 txCtx
                 & runExceptT . withExceptT
@@ -2221,7 +2221,7 @@ buildTransactionPure
     -> UTxOIndex WalletUTxO
     -> TransactionLayer k 'CredFromKeyK SealedTx
     -> ChangeAddressGen s
-    -> Read.ProtocolParameters era
+    -> Write.ProtocolParameters era
     -> PreSelection
     -> TransactionCtx
     -> ExceptT
@@ -2236,7 +2236,7 @@ buildTransactionPure
         withExceptT (Right . ErrConstructTxBody) . except $
             mkUnsignedTransaction txLayer @era
                 (unsafeShelleyOnlyGetRewardXPub (getState wallet))
-                (Read.pparamsWallet pparams)
+                (Write.pparamsWallet pparams)
                 txCtx
                 (Left preSelection)
 
@@ -2870,7 +2870,7 @@ delegationFee
 delegationFee db@DBLayer{atomically, walletsDB} netLayer
     txLayer ti era changeAddressGen walletId = do
     WriteTx.withRecentEra era $ \(recentEra :: WriteTx.RecentEra era) -> do
-        protocolParams <- Read.unsafeFromWalletProtocolParameters
+        protocolParams <- Write.unsafeFromWalletProtocolParameters
             <$> liftIO (currentProtocolParameters netLayer)
         wallet <- liftIO . atomically $ readDBVar walletsDB >>= \wallets ->
             case Map.lookup walletId wallets of
@@ -2884,7 +2884,7 @@ delegationFee db@DBLayer{atomically, walletsDB} netLayer
         unsignedTxBody <- wrapErrMkTransaction $
             mkUnsignedTransaction txLayer @era
                 (unsafeShelleyOnlyGetRewardXPub (getState wallet))
-                (Read.pparamsWallet protocolParams)
+                (Write.pparamsWallet protocolParams)
                 defaultTransactionCtx
                 -- It would seem that we should add a delegation action
                 -- to the partial tx we construct, this was not done
